backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
crispy-bootstrap4==2024.10
crispy-bootstrap5==2024.10
cryptography==44.0.0
dj-rest-auth==7.0.1
Django==5.1.3
django-allauth==65.2.0
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-oauth-toolkit==3.0.1
django-timezone==0.3
djangorestframework==3.15.2
djangorestframework-oauth==1.1.0
djangorestframework_simplejwt==5.4.0
freez==0.0.0
frozenlist==1.5.0
geoip2==4.8.1
git-filter-repo==2.47.0
idna==3.10
jwcrypto==1.5.6
maxminddb==2.6.2
multidict==6.1.0
mysqlclient==2.2.6
oauthlib==3.2.2
pillow==11.0.0
propcache==0.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
